FM1D2D-439: Changed NumberOfPeople and NumberOfUnits to a double. Former-commit-id: 30a4298cce5cb24667e26bd1bc697529a7989299 Former-commit-id: 95fb2c7c139676da4f4e358fed267b4aa9e43b18

+++ src/DeltaShell.Plugins.ImportExport.Sobek/PartialSobekImporter/SobekRRNwrwImporter.cs
@@ -430,7 +430,7 @@ namespace DeltaShell.Plugins.ImportExport.Sobek.PartialSobekImporter
 
         private void SetDryweatherFlows(NwrwData data, 
             Func<SobekRRNwrw, string> getDwaIdFunc,
-            Func<SobekRRNwrw, int> getNumberFunc,
+            Func<SobekRRNwrw, double> getNumberFunc,
             SobekRRNwrw dataToSetFrom,
             int index)
         {
